requirements.txt
telegram==0.0.1
python-telegram-bot==20.7
python-dotenv==1.0.0